Build cleanup
The build was a mess, now it is a little less of a mess. Top level
makefile handles depednecy between the src directory and the others as
well as setting build flags. Both the experiments and tests directories
link against libsecrecy.a instead of grabbing subsets of the C source.
The examples directory doesn't build anything because that code doesn't
compile as is and the fix is non-obvious.

Signed-off-by: Eric B Munson <munsoner@bu.edu>