Use TCP_NODELAY for our connections
A generic TCP socket uses Nagle's algorithm to improve transmission
efficieny by batching small packets into a single send. The
consequence of this for secrecy is that messages will wait until
a batching timeout occurs before sending. Because (at least for the
group-by experiment) all the messages are relatively small, this
adds significant, unnecessary waiting in communications. Instead,
we now set the TCP_NODELAY option on the communications sockets
(both send and receive). This cut runtimes by two orders of
magnitude for the smallest experiments.

Signed-off-by: Eric B Munson <munsoner@bu.edu>